+++ wavefront/data_source_derived_metrics.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/WavefrontHQ/go-wavefront-management-api"
+	"github.com/WavefrontHQ/go-wavefront-management-api/v2"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -174,10 +174,7 @@ func dataSourceDerivedMetricsRead(d *schema.ResourceData, m interface{}) error {
 	// Data Source ID is set to current time to always refresh
 	d.SetId(time.Now().UTC().String())
 
-	if err := d.Set(derivedMetricsKey, flattenDerivedMetrics(allDerivedMetrics)); err != nil {
-		return err
-	}
-	return nil
+	return d.Set(derivedMetricsKey, flattenDerivedMetrics(allDerivedMetrics))
 }
 
 func flattenDerivedMetrics(derivedMetrics []*wavefront.DerivedMetric) []map[string]interface{} {