+++ wavefront/data_source_roles.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/WavefrontHQ/go-wavefront-management-api"
+	"github.com/WavefrontHQ/go-wavefront-management-api/v2"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -48,10 +48,7 @@ func dataSourceRolesRead(d *schema.ResourceData, m interface{}) error {
 
 	// Data Source ID is set to current time to always refresh
 	d.SetId(time.Now().UTC().String())
-	if err := d.Set(rolesKey, flattenRoles(allRoles)); err != nil {
-		return err
-	}
-	return nil
+	return d.Set(rolesKey, flattenRoles(allRoles))
 }
 
 func flattenRoles(roles []*wavefront.Role) []map[string]interface{} {