+++ wavefront/data_source_users.go
@@ -3,7 +3,7 @@ package wavefront
 import (
 	"time"
 
-	"github.com/WavefrontHQ/go-wavefront-management-api"
+	"github.com/WavefrontHQ/go-wavefront-management-api/v2"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -63,10 +63,7 @@ func dataSourceUsersRead(d *schema.ResourceData, m interface{}) error {
 	}
 	// Data Source ID is set to current time to always refresh
 	d.SetId(time.Now().UTC().String())
-	if err := d.Set(usersKey, flattenUsers(users)); err != nil {
-		return err
-	}
-	return nil
+	return d.Set(usersKey, flattenUsers(users))
 }
 
 func flattenUsers(users []*wavefront.User) []map[string]interface{} {